+++ .babelrc.js
@@ -1,17 +1,43 @@
+/**
+ * This file is managed by backtrack
+ *
+ * source: @backtrack/preset-node-module
+ * namespace: babel
+ *
+ * DO NOT MODIFY
+ */
+
 'use strict';
+
+const Backtrack = require('@backtrack/core');
+
+const { configManager, pkg } = new Backtrack();
+
+const packageId = '@backtrack/preset-node-module';
 
 const babel = {
     presets: [
         [
-            'env',
+            pkg.resolve(packageId, 'babel-preset-env'),
             {
                 targets: {
                     node: '6.9.0',
                 },
+                useBuiltIns: true,
             },
         ],
     ],
-    plugins: ['transform-object-rest-spread'],
+    plugins: [
+        pkg.resolve(packageId, 'babel-plugin-dynamic-import-node'),
+        pkg.resolve(packageId, 'babel-plugin-transform-object-rest-spread'),
+        [
+            pkg.resolve(packageId, 'babel-plugin-transform-class-properties'),
+            { spec: true },
+        ],
+    ],
 };
 
-module.exports = babel;
+module.exports = configManager({
+    namespace: 'babel',
+    config: babel,
+});