+++ .prettierrc.js
@@ -1,11 +1,28 @@
+/**
+ * This file is managed by backtrack
+ *
+ * source: @backtrack/preset-style
+ * namespace: prettier
+ *
+ * DO NOT MODIFY
+ */
+
 'use strict';
 
-module.exports = {
+const Backtrack = require('@backtrack/core');
+
+const { configManager } = new Backtrack();
+
+const prettier = {
     semi: true,
     tabWidth: 4,
     singleQuote: true,
     trailingComma: 'all',
     arrowParens: 'always',
+
+    /**
+     * Needed for Node v6 support
+     */
     overrides: [
         {
             files: '*.js',
@@ -16,3 +33,8 @@ module.exports = {
         },
     ],
 };
+
+module.exports = configManager({
+    namespace: 'prettier',
+    config: prettier,
+});